fix: edit all game results, character statistics queries, change return type to single, change id to int

+++ app/src/main/java/com/example/typing_tracker/model/domain/Paragraph.kt
@@ -6,7 +6,7 @@ import com.example.typing_tracker.util.Difficulty
 
 @Entity(tableName = "PARAGRAPH_TABLE")
 data class Paragraph(
-    @PrimaryKey(autoGenerate = true) val id: Long,
+    @PrimaryKey(autoGenerate = true) val id: Int,
     val paragraph: String,
     val difficulty: Difficulty,
 )
